data_sources_processing_src/pyproject.toml
[tool.poetry]
name = "CPAoR Project"
version = "1.0.0"
description = ""
authors = ["CPAoR developers"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
nltk = "^3.8.1"
pandas = "^2.2.2"
pytesseract = "*"
python-dotenv = "1.0.1"
python_docx = "^1.1.0"
tqdm = "^4.66.2"
transformers = "^4.38.0"
openpyxl = "^3.1.2"
pymupdf = "^1.23.26"
openai = "^1.14.1"
beautifulsoup4 = "^4.12.3"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl"}
aiohttp = "3.10.5"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
